arePrivacySignalsApplicable
